Move ModalPopupService to Shared
