+++ src/app/components/main-components/main-components.module.ts
@@ -11,10 +11,10 @@ import {MatIconModule} from "@angular/material/icon";
 import {FormsModule} from "@angular/forms";
 import {MessagingDateFormatterPipe} from "../../pipes/messaging-date-formatter.pipe";
 import {MainComponent} from './main/main.component';
-import {SegmentedControlComponent} from "../shared/segmented-control/segmented-control.component";
+import {SegmentedControlComponent} from "../../shared/components/segmented-control/segmented-control.component";
 import {MainComponentsRoutingModule} from "./main-components-routing.module";
 import {MAT_DIALOG_DATA, MatDialogModule, MatDialogRef} from "@angular/material/dialog";
-import {ModalPopupService} from "./modal-popup.service";
+import {ModalPopupService} from "../../shared/services/modal-popup.service";
 import {MatFormFieldModule} from "@angular/material/form-field";
 import {MatSelectModule} from "@angular/material/select";
 import {MatInputModule} from "@angular/material/input";
@@ -22,11 +22,11 @@ import {IndividualUserDetailsPopupComponent} from './profile/individual-user-det
 import {MatDatepickerModule} from "@angular/material/datepicker";
 import {MatNativeDateModule} from "@angular/material/core";
 import {MatButtonModule} from "@angular/material/button";
-import {TextPairEditorComponent} from './helpers/text-pair-editor/text-pair-editor.component';
+import {TextPairEditorComponent} from '../../shared/components/text-pair-editor/text-pair-editor.component';
 import {EntityUserDetailsPopupComponent} from './profile/entity-user-details-popup/entity-user-details-popup.component';
 import {AddressAutoCompleterComponent} from './helpers/address-auto-completer/address-auto-completer.component';
 import {AddOpportunityPopupComponent} from './opportunities/add-opportunity-popup/add-opportunity-popup.component';
-import {RowTextEditorComponent} from './helpers/row-text-editor/row-text-editor.component';
+import {RowTextEditorComponent} from '../../shared/components/row-text-editor/row-text-editor.component';
 import {MatCheckboxModule} from "@angular/material/checkbox";
 import {GoogleMapsModule} from "@angular/google-maps";
 import {OpportunitiesFilterBarComponent} from './opportunities/opportunities-filter-bar/opportunities-filter-bar.component';
@@ -34,16 +34,16 @@ import {OpportunityDetailsComponent} from './opportunities/opportunity-details/o
 import {AddPostPopupComponent} from './news-center/add-post-popup/add-post-popup.component';
 import {PostSummaryCardComponent} from './profile/utility/post-summary-card/post-summary-card.component';
 import {PostDetailsComponent} from './news-center/post-details/post-details.component';
-import {PaginatorComponent} from '../shared/paginator/paginator.component';
+import {PaginatorComponent} from '../../shared/components/paginator/paginator.component';
 import {NgxSpinnerModule} from "ngx-spinner";
 import {MatTooltipModule} from "@angular/material/tooltip";
-import {HorizontalScrollableTabsComponent} from "../shared/horizontal-scrollable-tabs/horizontal-scrollable-tabs.component";
+import {HorizontalScrollableTabsComponent} from "../../shared/components/horizontal-scrollable-tabs/horizontal-scrollable-tabs.component";
 import {ProfilePostsComponent} from './profile/profile-posts/profile-posts.component';
 import {ProfileAboutComponent} from './profile/profile-about/profile-about.component';
 import {ProfileOpportunitiesComponent} from './profile/profile-opportunities/profile-opportunities.component';
 import {MatGridListModule} from "@angular/material/grid-list";
-import { ModalPopupComponent } from './helpers/modal-popup/modal-popup.component';
-import {SharedModule} from "../shared/shared.module";
+import { ModalPopupComponent } from '../../shared/components/modal-popup/modal-popup.component';
+import {SharedModule} from "../../shared/shared.module";
 
 @NgModule({
     declarations: [
@@ -56,11 +56,9 @@ import {SharedModule} from "../shared/shared.module";
         MessagingDateFormatterPipe,
         MainComponent,
         IndividualUserDetailsPopupComponent,
-        TextPairEditorComponent,
         EntityUserDetailsPopupComponent,
         AddressAutoCompleterComponent,
         AddOpportunityPopupComponent,
-        RowTextEditorComponent,
         OpportunitiesFilterBarComponent,
         OpportunityDetailsComponent,
         AddPostPopupComponent,
@@ -68,8 +66,7 @@ import {SharedModule} from "../shared/shared.module";
         PostDetailsComponent,
         ProfilePostsComponent,
         ProfileAboutComponent,
-        ProfileOpportunitiesComponent,
-        ModalPopupComponent
+        ProfileOpportunitiesComponent
     ],
     imports: [
         CommonModule,
@@ -93,17 +90,6 @@ import {SharedModule} from "../shared/shared.module";
         GoogleMapsModule,
         // Misc
         NgxSpinnerModule
-    ],
-    providers: [
-        {
-            provide: MatDialogRef,
-            useValue: {}
-        },
-        {
-            provide: MAT_DIALOG_DATA,
-            useValue: {}
-        },
-        ModalPopupService
     ]
 })
 export class MainComponentsModule {
