+++ src/app/components/main-components/news-center/NewsCenterViewModel.ts
@@ -6,7 +6,7 @@ import {exhaustMap, filter, map, share} from 'rxjs/operators';
 import {Post, UsersAndPosts} from '../../../models/model classes/posts/Post';
 import {PageData, PageDataMetadata} from '../../../models/model classes/common/PageData';
 import {User} from '../../../models/model classes/user/User';
-import {ModalPopupService} from "../modal-popup.service";
+import {ModalPopupService} from "../../../shared/services/modal-popup.service";
 
 export class NewsCenterViewModel {
     pageChanged$: Subject<number> = new Subject();
