+++ src/app/components/main-components/news-center/news-center.component.ts
@@ -7,10 +7,10 @@ import {Post, UsersAndPosts} from '../../../models/model classes/posts/Post';
 import {User} from '../../../models/model classes/user/User';
 import {catchError, map, mergeMap, takeUntil} from 'rxjs/operators';
 import {Title} from '@angular/platform-browser';
-import {PaginatorComponent} from '../../shared/paginator/paginator.component';
+import {PaginatorComponent} from '../../../shared/components/paginator/paginator.component';
 import {of, Subject} from 'rxjs';
 import {NgxSpinnerService} from 'ngx-spinner';
-import {ModalPopupService} from "../modal-popup.service";
+import {ModalPopupService} from "../../../shared/services/modal-popup.service";
 
 @Component({
     selector: 'app-news-center',
