+++ src/app/components/main-components/profile/profile-about/ProfileAboutViewModel.ts
@@ -6,7 +6,7 @@ import {entityUserTypes, individualUserTypes} from "../../../../models/BusinessC
 import {IndividualUserDetailsPopupComponent} from "../individual-user-details-popup/individual-user-details-popup.component";
 import {EntityUserDetailsPopupComponent} from "../entity-user-details-popup/entity-user-details-popup.component";
 import {filter, map, mergeMap, take} from "rxjs/operators";
-import {ModalPopupService} from "../../modal-popup.service";
+import {ModalPopupService} from "../../../../shared/services/modal-popup.service";
 
 export class ProfileAboutViewModel {
     constructor(private usersApi: UserApiService, private modalPopupService: ModalPopupService) {
