+++ src/app/components/main-components/profile/profile.component.ts
@@ -3,7 +3,7 @@ import {UserApiService} from "../../../services/user-api.service";
 import {User} from "../../../models/model classes/user/User";
 import {Router} from "@angular/router";
 import {UserDetails} from "../../../models/model classes/user/UserDetails";
-import {ModalPopupService} from "../modal-popup.service";
+import {ModalPopupService} from "../../../shared/services/modal-popup.service";
 import {PostsService} from "../../../services/posts.service";
 import {Post} from "../../../models/model classes/posts/Post";
 import {ProfileViewModel} from "./ProfileViewModel";
