+++ src/app/shared/components/horizontal-scrollable-tabs/horizontal-scrollable-tabs.component.ts
