+++ src/app/shared/components/paginator/paginator.component.scss
