+++ src/app/shared/components/paginator/paginator.component.spec.ts
