+++ src/app/shared/components/row-text-editor/row-text-editor.component.html
