+++ src/app/shared/components/segmented-control/segmented-control.component.scss
