+++ src/app/shared/components/segmented-control/segmented-control.component.ts
