+++ src/app/shared/components/text-pair-editor/text-pair-editor.component.spec.ts
