+++ src/app/shared/shared.module.ts
@@ -10,18 +10,26 @@ import {MatToolbarModule} from "@angular/material/toolbar";
 import {MatIconModule} from "@angular/material/icon";
 import {MatButtonModule} from "@angular/material/button";
 import {MatCheckboxModule} from "@angular/material/checkbox";
-import {HorizontalScrollableTabsComponent} from "./horizontal-scrollable-tabs/horizontal-scrollable-tabs.component";
-import {PaginatorComponent} from "./paginator/paginator.component";
-import {SegmentedControlComponent} from "./segmented-control/segmented-control.component";
+import {HorizontalScrollableTabsComponent} from "./components/horizontal-scrollable-tabs/horizontal-scrollable-tabs.component";
+import {PaginatorComponent} from "./components/paginator/paginator.component";
+import {SegmentedControlComponent} from "./components/segmented-control/segmented-control.component";
+import {TextPairEditorComponent} from "./components/text-pair-editor/text-pair-editor.component";
+import {RowTextEditorComponent} from "./components/row-text-editor/row-text-editor.component";
+import {ModalPopupComponent} from "./components/modal-popup/modal-popup.component";
+import {FormsModule} from "@angular/forms";
 
 @NgModule({
     declarations: [
         HorizontalScrollableTabsComponent,
         PaginatorComponent,
-        SegmentedControlComponent
+        SegmentedControlComponent,
+        TextPairEditorComponent,
+        RowTextEditorComponent,
+        ModalPopupComponent
     ],
     imports: [
         CommonModule,
+        FormsModule,
         MatDialogModule,
         MatFormFieldModule,
         MatSelectModule,
@@ -36,7 +44,10 @@ import {SegmentedControlComponent} from "./segmented-control/segmented-control.c
     exports: [
         HorizontalScrollableTabsComponent,
         PaginatorComponent,
-        SegmentedControlComponent
+        SegmentedControlComponent,
+        TextPairEditorComponent,
+        RowTextEditorComponent,
+        ModalPopupComponent
     ]
 })
 export class SharedModule {
